backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "backend-app"
version = "0.1.0"
description = "Backend service for the Spongebob project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "pydantic-settings>=2.3.4",
    "sse-starlette>=2.1.0",
    "python-multipart>=0.0.9",
    "anthropic>=0.29.0",
    "pdfminer.six>=20240705",
    "mcp>=1.6.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "rapidfuzz>=3.0.0",
]

[tool.uv]
dev-dependencies = [
    "requests>=2.31.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"] 